[#55883844] Add option to optmize rotated indices
Once an index has been rotated out it will no longer be hot and live
in mem. We want to be able to optimize the old indices such that we
can get as fast as possible access when cold.

This functionality can be enabled by adding the argument --optimize-old
to es-rotate, with --optimize-maxage expressing the cutoff time delta
(in days) before they get optimized (default 1).